feat: Implement BT v4 closed-loop delivery system

+++ ros2_ws/src/uav_bt_agent/include/uav_bt_agent/plugins/action/plan_path_action.hpp
@@ -19,9 +19,9 @@ public:
     static BT::PortsList providedPorts()
     {
         return {
-            BT::InputPort<double>("goal_x"),
-            BT::InputPort<double>("goal_y"),
-            BT::InputPort<double>("goal_z"),
+            BT::InputPort<float>("goal_x"),
+            BT::InputPort<float>("goal_y"),
+            BT::InputPort<float>("goal_z"),
             BT::OutputPort<std::vector<geometry_msgs::msg::Point>>("waypoints")
         };
     }
